replace old argument by the observer struct
